Add Janitor core component
