+++ lib/api/controllers/adminController.js
@@ -24,20 +24,12 @@
 const
   Bluebird = require('bluebird'),
   {
-    InternalError: KuzzleInternalError,
     BadRequestError
   } = require('kuzzle-common-objects').errors,
   Request = require('kuzzle-common-objects').Request,
   {
     assertArgsHasAttribute
-  } = require('../../util/requestAssertions'),
-  path = require('path'),
-  fs = require('fs-extra'),
-  os = require('os'),
-  moment = require('moment'),
-  dumpme = require('dumpme'),
-  zlib = require('zlib'),
-  glob = require('glob');
+  } = require('../../util/requestAssertions');
 
 let
   _dump = false,
@@ -132,9 +124,9 @@ class AdminController {
       throw new BadRequestError('Dump generation is disabled.');
     }
 
-    const suffix = request.input.args.suffix ? `-${request.input.args.suffix}` : '';
+    const suffix = request.input.args.suffix ? request.input.args.suffix : '';
 
-    generateDump(this.kuzzle, suffix);
+    this.kuzzle.janitor.dump(suffix);
 
     return Bluebird.resolve({ acknowledge: true });
   }
@@ -154,227 +146,6 @@ class AdminController {
     return Bluebird.resolve({ acknowledge: true });
   }
 
-}
-
-/**
- * Create a dump
- *
- * @param {Kuzzle} kuzzle
- * @param {string} suffix
- * @returns {Promise}
- */
-/* eslint-disable no-console */
-function generateDump (kuzzle, suffix) {
-  let dumpPath;
-  _dump = true;
-
-  return new Bluebird((resolve, reject) => {
-    dumpPath = path.join(
-      path.normalize(kuzzle.config.dump.path),
-      moment().format(kuzzle.config.dump.dateFormat).concat(suffix).substring(0, 200)
-    );
-
-    console.log('===========================================================================');
-    cleanUpHistory(kuzzle);
-
-    console.log(`Generating dump in ${dumpPath}`);
-    try {
-      fs.mkdirsSync(dumpPath);
-    }
-    catch (e) {
-      if (e.message.startsWith('EEXIST')) {
-        console.log('Dump directory already exists. Skipping..');
-        return reject('Dump directory already exists. Skipping..');
-      }
-
-      console.log('ERROR: unknown error while trying to create dump folder');
-      return reject(`ERROR: unknown error while trying to create dump folder: ${e.message}`);
-    }
-
-    // dumping kuzzle configuration
-    console.log('> dumping kuzzle configuration');
-    fs.writeFileSync(path.join(dumpPath, 'config.json'), JSON.stringify(kuzzle.config, null, ' ').concat('\n'));
-
-    // dumping plugins configuration
-    console.log('> dumping plugins configuration');
-    fs.writeFileSync(path.join(dumpPath, 'plugins.json'), JSON.stringify(kuzzle.pluginsManager.getPluginsDescription(), null, ' ').concat('\n'));
-
-    // dumping nodejs configuration
-    console.log('> dumping nodejs configuration');
-    fs.writeFileSync(path.join(dumpPath, 'nodejs.json'), JSON.stringify({
-      env: process.env,
-      config: process.config,
-      argv: process.argv,
-      versions: process.versions,
-      release: process.release,
-      moduleLoadList: process.moduleLoadList
-    }, null, ' ').concat('\n'));
-
-    // dumping os configuration
-    console.log('> dumping os configuration');
-    fs.writeFileSync(path.join(dumpPath, 'os.json'), JSON.stringify({
-      platform: os.platform(),
-      loadavg: os.loadavg(),
-      uptime: os.uptime(),
-      cpus: os.cpus(),
-      mem: {
-        total: os.totalmem(),
-        free: os.freemem()
-      },
-      networkInterfaces: os.networkInterfaces()
-    }, null, ' ').concat('\n'));
-
-    // dumping std err/out
-    console.log('> dumping std err/out');
-    if (!process.env.pm_err_log_path && !process.env.pm_log_path) {
-      console.log('... no log found');
-      return resolve();
-    }
-
-    const logDir = path.dirname(process.env.pm_err_log_path || process.env.pm_log_path);
-    fs.readdir(logDir, (err, files) => {
-      if (err) {
-        return reject(new KuzzleInternalError('... error reading log directory'));
-      }
-
-      Bluebird.all(files.map(file => Bluebird.promisify(fs.stat)(path.join(logDir, file))
-        .then(stats => {
-          return {
-            path: path.join(logDir, file),
-            stats
-          };
-        }))
-      )
-        .then(fileStats => {
-          const keep = {};
-
-          for (const statFile of fileStats) {
-            const root = path.basename(statFile.path).replace(/(-\d+)?\.[^.]+$/, '');
-            if (!keep[root] || keep[root].stats.ctime < statFile.stats.ctime) {
-              keep[root] = statFile;
-            }
-          }
-
-          const promises = [];
-          for (const root of Object.keys(keep)) {
-            const statFile = keep[root];
-
-            fs.ensureDirSync(path.join(dumpPath, 'logs'));
-            // eslint-disable-next-line no-loop-func
-            promises.push(new Bluebird(res => {
-              fs.createReadStream(statFile.path)
-                .pipe(zlib.createGzip())
-                .pipe(fs.createWriteStream(path.join(dumpPath, 'logs', root + '.gz')))
-                .on('finish', () => {
-                  res();
-                });
-            }));
-          }
-          return Bluebird.all(promises);
-        })
-        .then(() => {
-          resolve();
-        })
-        .catch(e => {
-          reject(`fileStats: ${e.message}`);
-        });
-    });
-  })
-    .then(() => new Bluebird((resolve, reject) => {
-      // core-dump
-      console.log('> generating core-dump');
-      dumpme(kuzzle.config.dump.gcore || 'gcore', `${dumpPath}/core`);
-
-      // Gzip the core
-      glob(`${dumpPath}/core*`, (err, res) => {
-        if (err) {
-          console.log(err);
-          return reject(err);
-        }
-
-        if (!res[0]) {
-          console.log('> warning: could not generate dump');
-          return resolve();
-        }
-
-        const readStream = fs.createReadStream(res[0]);
-        const writeStream = fs.createWriteStream(`${dumpPath}/core.gz`);
-
-        readStream
-          .pipe(zlib.createGzip())
-          .pipe(writeStream)
-          .on('finish', () => {
-            // rm the original core file
-            try {
-              fs.unlinkSync(res[0]);
-              resolve();
-            }
-            catch (e) {
-              resolve(e.message);
-            }
-          });
-      });
-    }))
-    .then(() => {
-      // copy node binary
-      console.log('> copy node binary');
-      fs.copySync(process.execPath, path.join(dumpPath, 'node'));
-
-      // dumping Kuzzle's stats
-      console.log('> dumping kuzzle\'s stats');
-      return kuzzle.statistics.getAllStats(new Request({action: 'getAllStats', controller: 'statistics'}));
-    })
-    .then(response => {
-      fs.writeFileSync(path.join(dumpPath, 'statistics.json'), JSON.stringify(response.hits, null, ' ').concat('\n'));
-
-      console.log('Done.');
-      console.log('[ℹ] You can send the folder to the kuzzle core team at support@kuzzle.io');
-      console.log('===========================================================================');
-    })
-    .then(() => dumpPath)
-    .finally(() => {
-      _dump = false;
-    });
-}
-
-function cleanUpHistory(kuzzle) {
-  const
-    config = kuzzle.config.dump,
-    dumpPath = path.normalize(kuzzle.config.dump.path);
-
-  try {
-    fs.accessSync(dumpPath, fs.constants.R_OK | fs.constants.W_OK | fs.constants.X_OK);
-  }
-  catch(e) {
-    return;
-  }
-
-  const dumps = fs.readdirSync(dumpPath)
-    .map(file => {
-      const filepath = `${dumpPath}/${file}`;
-      return {path: filepath, stat: fs.statSync(filepath)};
-    })
-    .filter(prop => prop.stat.isDirectory())
-    .sort((a, b) => {
-      if (a.stat.birthtime.getTime() === b.stat.birthtime.getTime()) {
-        return 0;
-      }
-
-      return a.stat.birthtime < b.stat.birthtime ? -1 : 1;
-    });
-
-  while (dumps.length >= config.history.reports) {
-    const dir = dumps.shift().path;
-
-    fs.removeSync(dir);
-  }
-
-  for (let i = 0; i < dumps.length - config.history.coredump; i++) {
-    const corefiles = glob.sync(`${path.normalize(dumps[i].path)}/core*`);
-    if (corefiles[0]) {
-      fs.unlinkSync(corefiles[0]);
-    }
-  }
 }
 
 module.exports = AdminController;
