+++ lib/api/kuzzle.js
@@ -32,6 +32,7 @@ const
   EntryPoints = require('./core/entrypoints'),
   FunnelController = require('./controllers/funnelController'),
   HotelClerk = require('./core/hotelClerk'),
+  Janitor = require('./core/janitor'),
   IndexCache = require('./core/indexCache'),
   InternalEngine = require('../services/internalEngine'),
   Notifier = require('./core/notifier'),
@@ -101,6 +102,9 @@ class Kuzzle extends EventEmitter {
 
     // Validation core component
     this.validation = new Validation(this);
+
+    // Janitor component
+    this.janitor = new Janitor(this);
   }
 
   /**
